+++ ChangeLog.md
@@ -1,7 +1,12 @@
-  Version 1.0.0
+Version 1.0
   - First public release
 
-  Version 1.1.0
+Version 1.1
   - minor "Cache-Control" parameters order changes
   - added bpoDelegateIndexToInheritedDefault to delegate index.html to Default()
   - added bpoDelegate404ToInherited_404 to delegate 404.html to _404()
+
+Version 1.2
+  - fix "Accept-Encoding" parsing when gzip in the end of encodings list
+  - make Pre-Build events notice more visible in tests and demo
+  - minor code refactoring